requirements.txt
aiohttp[speedup]>=3.7.4,<4
tzdata==2022.7,>=2021.10
orjson
python-magic==0.4.13
redis>=4.2.0rc1
discord.py==2.2.0
pandas==1.3.5
coloredlogs==15.0.1
git+https://github.com/gorialis/jishaku
pydantic[dotenv]>=1.10.4
git+https://github.com/bryanforbes/asyncpg-stubs
Pillow>=9.4.0
asyncpg>=0.27.0
emoji==1.7.0
pydantic[dotenv]>=1.10.4
